requirements.txt
openai
requests
moviepi
google-cloud-storage